+++ packages/inqvine_core_firebase/lib/inqvine_core_firebase.dart
@@ -1,7 +1,12 @@
 library inqvine_core_firebase;
 
+export 'proto/application_status.pb.dart';
+export 'proto/application_status.pbenum.dart';
+export 'proto/application_status.pbjson.dart';
+
 export 'services/inqvine_auth_service.dart';
 export 'services/inqvine_admin_service.dart';
+export 'services/inqvine_system_service.dart';
 
 export 'package:firebase_auth/firebase_auth.dart';
 export 'package:cloud_firestore/cloud_firestore.dart';
@@ -12,6 +17,7 @@ import 'package:firebase_core/firebase_core.dart';
 
 import 'package:inqvine_core_firebase/services/inqvine_admin_service.dart';
 import 'package:inqvine_core_firebase/services/inqvine_auth_service.dart';
+import 'package:inqvine_core_firebase/services/inqvine_system_service.dart';
 import 'package:inqvine_core_main/inqvine_core_main.dart';
 
 InqvineCoreFirebase get inqvineFirebase => InqvineCoreFirebase.instance;
@@ -30,6 +36,7 @@ class InqvineCoreFirebase {
 
     await inqvine.registerService<InqvineAuthService>(InqvineAuthService.instance);
     await inqvine.registerService<InqvineAdminService>(InqvineAdminService.instance);
+    await inqvine.registerService<InqvineSystemService>(InqvineSystemService.instance);
   }
 
   FirebaseAuth get firebaseAuth => inqvine.getFromLocator<FirebaseAuth>();
@@ -37,4 +44,5 @@ class InqvineCoreFirebase {
 
   InqvineAuthService get authService => inqvine.getFromLocator<InqvineAuthService>();
   InqvineAdminService get adminService => inqvine.getFromLocator<InqvineAdminService>();
+  InqvineSystemService get systemService => inqvine.getFromLocator<InqvineSystemService>();
 }