+++ packages/inqvine_core_ui/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.0.12-nullsafety.0] - Add System Service into Firebase Core
+
+* Handle automatically preventing using of the application with downtimeBuilder and applicationStatus
+
 ## [0.0.11-nullsafety.0] - Add UI package
 
 * Common UI helpers in inqvine_core_ui